Update yarn installation instructions.
When trying to run the command to install `vuex-electron` via yarn, you'll get an error in the terminal now since yarn as replaced `install` with the `add` keyword. 

Example of the error returned is as follows:
```
error `install` has been replaced with `add` to add new dependencies. Run "yarn add vuex-electron" instead.
```